We can now register and login but read description
Register and login operations work but still some work needs to be done. We have to register and login according to the account type. Not sure about the middleware, I did not test it. Also, instead of simplejwt I decided to use pyjwt. It is easier to use for our purposes. I did the necessary changes in requirements.txt. Finally I did not check update profile etc. I will look at them and also I changed the password max length in init.sql because hashed password does not fit to a varchar(50) field. But right now I did not use hashed password which needs to be done as well.